runtime model generation

+++ data_objects/parser_inner_object/Vertex.h
@@ -10,12 +10,12 @@
 
 class Vertex {
 public:
-    Vertex(uint16_t x, uint16_t y, uint16_t z) : x(x), y(y), z(z) {}
+    Vertex(float x, float y, float z) : x(x), y(y), z(z) {}
 
 public:
-    uint16_t x;
-    uint16_t y;
-    uint16_t z;
+    float x;
+    float y;
+    float z;
 
 };
